Remove LocalstackS3WaveformConsumer
