+++ pom.xml
@@ -260,8 +260,10 @@
           </systemPropertyVariables>
           <environmentVariables>
             <AUDIOWAVEFORM_S3_BUCKET>test-audiowaveform-resources</AUDIOWAVEFORM_S3_BUCKET>
+            <AWS_ACCESS_KEY_ID>test</AWS_ACCESS_KEY_ID><!-- access key and secret key must be set (to anything) -->
             <AWS_ENDPOINT_URL>http://localhost:${localstack.edge.port}</AWS_ENDPOINT_URL>
             <AWS_DEFAULT_REGION>us-east-1</AWS_DEFAULT_REGION>
+            <AWS_SECRET_ACCESS_KEY>test</AWS_SECRET_ACCESS_KEY>
           </environmentVariables>
           <forkCount>1</forkCount>
           <argLine>${jacoco.agent.arg}</argLine>
@@ -276,7 +278,7 @@
           </systemPropertyVariables>
           <environmentVariables>
             <AUDIOWAVEFORM_S3_BUCKET>test-audiowaveform-resources</AUDIOWAVEFORM_S3_BUCKET>
-            <AWS_ACCESS_KEY_ID>test</AWS_ACCESS_KEY_ID><!-- access key and secret key must be set (to anything) -->
+            <AWS_ACCESS_KEY_ID>test</AWS_ACCESS_KEY_ID>
             <AWS_ENDPOINT_URL>http://localhost:${localstack.edge.port}</AWS_ENDPOINT_URL>
             <AWS_DEFAULT_REGION>us-east-1</AWS_DEFAULT_REGION>
             <AWS_SECRET_ACCESS_KEY>test</AWS_SECRET_ACCESS_KEY>
